Move BIR tests to its own class
